coproc/tests: Seperating out fixture duties
- Makes for more composable test fixtures, preparing the groundwork for a
fixture that won't inherit from coproc_test_fixture but will be able to use the
supervisor service as well
