+++ src/v/coproc/CMakeLists.txt
@@ -36,7 +36,7 @@ rp_test(
   SOURCES
     tests/script_manager_service_test.cc
     tests/supervisor.cc
-    tests/coproc_fixture_base.cc
+    tests/coproc_test_fixture.cc
     tests/router_test.cc
     tests/utils.cc
   DEFINITIONS BOOST_TEST_DYN_LINK
